Implement Skin loader.
Need to separate this into multiple files soon, but I at least want to get everything working first. This implements skin loading, compartmentalized to a new file called `PN_LoadSkins.php`. Most of the work is done outside of this script; the biggest thing here is that we have all of our skins and extensions on GitHub synced to the version we want to deploy. If we ever want to synchronously update everything we'll probably want to script updating the `puyonexus` branches, which is pretty easy to do (truth be told, I scripted it just now for sake of making the branches.)